Re-coded the testing function
The previous tests were deleted; they didn't make sense with the fundamental restructuring of the graphical methods. Instead, it now includes a few tests of the qualitative behavior of our numerical simulations. I plan to add a quantitative facet once I understand implementation of the Wasserstein metric on periodic domains.